pyproject.toml
[tool.poetry]
name = "homework-01"
version = "0.1.0"
description = ""
authors = ["Aleksandr Akimenko <alexandr.akimenko@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
pyyaml = "^6.0.2"
structlog = "^24.4.0"
pytest = "^8.3.2"
types-pyyaml = "^6.0.12.20240808"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.8.0"
types-pyyaml = "^6.0.12.20240808"
black = "^24.8.0"
isort = "^5.13.2"
mypy = "^1.5.1"
flake8 = "^7.1.1"
flake8-bugbear = "^23.3.21"
flake8-import-order = "^0.18.1"
ruff = "^0.6.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.isort]
profile = "black"

[tool.flake8]
max-line-length = 120

[tool.mypy]
ignore_missing_imports = true

[tool.mypy.overrides]
module = "yaml"
ignore_missing_imports = true